==> Documents/2.Cleaning/3. Adding Players + Transfers Data.ipynb <==
import pandas as pd 
import numpy as np
import regex as re

#### Adding The Player Info

players_2020 = pd.read_csv("C:/Users/Titoran/Documents/Cleaned Players Data/Players_2020.csv", index_col = [0])
players_2021 = pd.read_csv("C:/Users/Titoran/Documents/Cleaned Players Data/Players_2021.csv", index_col = [0])
players_2022 = pd.read_csv("C:/Users/Titoran/Documents/Cleaned Players Data/Players_2022.csv", index_col = [0])

df_NER = pd.read_csv('Rumours_Sent.csv', index_col = [0] )

df_NER = df_NER.loc[df_NER["Score"] == "Positive",]
df_NER = df_NER.drop(['Score'], axis=1)

#Unlisting the players to then join them
#df_NER["Player"] = [str(''.join(map(str, l))) for l in df_NER['Player']]

#Maing sure just the player name is in the column
df_NER["Player"] = df_NER["Player"].apply(lambda x: x[2:-2])

df2020 = df_NER.loc[df_NER["Year"] == 2020,]

df2021 = df_NER.loc[df_NER["Year"] == 2021,]

df2022 = df_NER.loc[df_NER["Year"] == 2022,]

df0 = df2020.merge(players_2020, how='left', on='Player')

df0 = df0.drop(['Year_y'], axis=1)
df0 = df0.rename(columns={'Year_x': 'Year'})

df1 = df2021.merge(players_2021, how='left', on='Player')

df1 = df1.drop(['Year_y'], axis=1)
df1 = df1.rename(columns={'Year_x': 'Year'})

df2 = df2022.merge(players_2022, how='left', on='Player')

df2 = df2.drop(['Year_y'], axis=1)
df2 = df2.rename(columns={'Year_x': 'Year'})

df_players = pd.concat([ df0, df1, df2], axis=0)

#### Club Name Cleaning

In the dictionary of clubs for NER, I added common nicknames of clubs which could be used. This section retransforms these names to the clubs proper name

df_players["Clubs"] = [l.replace("PSG", "Paris Saint-Germain") for l in df_players["Clubs"]]

df_players["Clubs"] = [l.replace("Reds", "Manchester United") for l in df_players["Clubs"]]
df_players["Clubs"] = [l.replace("Red Devils", "Manchester United") for l in df_players["Clubs"]]

df_players["Clubs"] = [l.replace("Blues", "Chelsea") for l in df_players["Clubs"]]

df_players["Clubs"] = [l.replace("Gunners", "Arsenal") for l in df_players["Clubs"]]

df_players["Clubs"] = [l.replace("Spurs", "Tottenham") for l in df_players["Clubs"]]

#### Finding the interested club

#Similiarty between words function to account for small differences in how cubs are spelt
from difflib import SequenceMatcher
def similar(a, b):
    return SequenceMatcher(None, a, b).ratio()

#This allows to iterate through a list
from ast import literal_eval

#For each club in the clubs column, we check if the name is, or is similar to the name of the players club.
#If its not , we assume this club is interested in buying the player and add it to a new column which will contain the interested clubs.
n = 0
interested_club = []
for row in df_players["Clubs"] : 
    temp = []
    for club in literal_eval(row) : 
        if (similar(club, str(df_players.iloc[n,12])) < 0.7) & (club not in str(df_players.iloc[n,12]))  :
            temp.append(club)
    interested_club.append(temp)
    n += 1

df_players["Interested_clubs"] = interested_club

#df_players.iloc[0:50,[3,4,12,13]]

#### Checking for the Transfer

transfers = pd.read_csv("C:/Users/Titoran/Documents/Transfers Data/transfers.csv", index_col = [0])

#For each row, we check if the player has been involved in a transfer to one of the interested clubs. If yes, we consider the outcome to be True

Outcome = []
#n allows to keep track of which row we are iterating through
n = 0 
for player in df_players["Player"] :
    temp = 0 
    try : 
        for interested in df_players.iloc[n,13] : 
            for club in transfers.loc[(transfers["player_name"] == player), "club_name"] :
                if ((similar(club, str(interested)) > 0.7) | (club in str(interested)) | (str(interested) in club)):  
                    temp = 1
        if temp == 1 : 
            Outcome.append("True")
        else : 
            Outcome.append("False")
    except IndexError : 
            Outcome.append("False")
    n += 1

df_players["Outcome"] = Outcome

df_players

#Checking the transfers which did happen
#df_players.loc[df_players["Outcome"] == "True",]

df_players.to_csv('Uncleaned_Data.csv', encoding='utf-8')
