==> Documents/2.Cleaning/4. Outlets Cleaning.ipynb <==
import pandas as pd 
import numpy as np
import regex as re

df_players = pd.read_csv("Uncleaned_Data.csv", index_col = [0] )

len(df_players['Outlet'].unique())

#### Outlet Cleaning

#Removes trailing spaces
df_players = df_players.applymap(lambda x: x.strip() if isinstance(x, str) else x)

#Sometimes the outlet says outlet x via outlet y meaning y reported on it but x said it. Therefore we only want to keep outlet xx.
for outlet in df_players.loc[df_players['Outlet'].str.contains("via"), "Outlet"] : 
    df_players.loc[df_players['Outlet'] == outlet, "Outlet"] = outlet.split("via")[0]
    
for outlet in df_players.loc[df_players['Outlet'].str.contains("on Sunday"), "Outlet"] : 
    df_players.loc[df_players['Outlet'] == outlet, "Outlet"] = outlet.split("on Sunday")[0]

#Cleaning the outlet text
df_players['Outlet'] = [str(w).replace('external', '') for w in df_players['Outlet']]

df_players.loc[df_players['Outlet'].str.contains("Fabrizio Romano"), "Outlet"] = "Fabrizio Romano"

df_players.loc[df_players['Outlet'].str.contains("Mail Online"), "Outlet"] = "Mail"
df_players.loc[df_players['Outlet'].str.contains("MailOnline"), "Outlet"] = "Mail" 
df_players.loc[df_players['Outlet'].str.contains("Daily Mail"), "Outlet"] = "Mail" 
df_players.loc[df_players['Outlet'].str.contains("Mail"), "Outlet"] = "Mail" 

df_players.loc[df_players['Outlet'].str.contains("Sun"), "Outlet"] = "The Sun" 

df_players.loc[df_players['Outlet'].str.contains("Give Me Sport"), "Outlet"] = "Sport"

df_players.loc[df_players['Outlet'].str.contains("Chronicle"), "Outlet"] = "Chronicle" 

df_players.loc[df_players['Outlet'].str.contains("Star"), "Outlet"] = "Star"

df_players.loc[df_players['Outlet'].str.contains("90"), "Outlet"] = "90 Min"

df_players.loc[df_players['Outlet'].str.contains("Sky"), "Outlet"] = "Sky Sports"

df_players.loc[df_players['Outlet'].str.contains("Sunday Mirror"), "Outlet"] = "Mirror" 

df_players.loc[df_players['Outlet'].str.contains("talkSPORT"), "Outlet"] = "TalkSPORT" 
df_players.loc[df_players['Outlet'].str.contains("TalkSPORT"), "Outlet"] = "TalkSport" 

df_players.loc[df_players['Outlet'].str.contains("Express"), "Outlet"] = "Express"

df_players.loc[df_players['Outlet'].str.contains("Athletic"), "Outlet"] = "The Athletic"

df_players.loc[df_players['Outlet'].str.contains("Twitter"), "Outlet"] = "Independent Journalists" 
df_players.loc[df_players['Outlet'].str.contains("Nicolo Schira"), "Outlet"] = "Independent Journalists" 
df_players.loc[df_players['Outlet'].str.contains("Sport Witness"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Transfer Window Podcast"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Gianluigi Longari"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Gianluca"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Florian Plettenberg"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Mit Geredet"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Gokhan Dinc"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Marco Barzaghi"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Ruhr Nachrichten"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Lucas Navarrete"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Alan Nixon"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Tancredi Palmeri"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Gianluigi Longari"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Independent"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Gerard Romero"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Sebastian Giovanelli"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Fabrice Hawkins"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Algemeen Dagblad"), "Outlet"] = "Independent Journalists"
df_players.loc[df_players['Outlet'].str.contains("Jacque Talbot"), "Outlet"] = "Independent Journalists"

df_players.loc[df_players['Outlet'].str.contains("Standard"), "Outlet"] = "Evening Standard"

df_players.loc[df_players['Outlet'].str.contains("Marca"), "Outlet"] = "Marca"

df_players.loc[df_players['Outlet'].str.contains("Mundo Deportivo"), "Outlet"] = "Mundo Deportivo" 

df_players.loc[df_players['Outlet'].str.contains("CalcioMercato"), "Outlet"] = "Calcio mercato" 
df_players.loc[df_players['Outlet'].str.contains("Calciomercato"), "Outlet"] = "Calcio mercato" 

df_players.loc[df_players['Outlet'].str.contains("Corriere"), "Outlet"] = "Corriere dello Sport" "

df_players.loc[df_players['Outlet'].str.contains("Bild"), "Outlet"] = "Bild"

df_players = df_players.applymap(lambda x: x.strip() if isinstance(x, str) else x)

#### Special Cases

#LAST CHECK
for outlet in df_players.loc[df_players['Outlet'].str.contains("-"), "Outlet"] : 
    df_players.loc[df_players['Outlet'] == outlet, "Outlet"] = outlet.split("-")[0]

for outlet in df_players.loc[df_players['Outlet'].str.contains(","), "Outlet"] : 
    df_players.loc[df_players['Outlet'] == outlet, "Outlet"] = outlet.split(",")[0]

#Saving
df_players.to_csv("Outlets_Cleaned.csv", encoding='utf-8')
